third_party/WebKit/LayoutTests/editing/execCommand/nsresponder-indent.html

<!doctype html>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="../assert_selection.js"></script>
<script>
test(() => assert_selection(
    '<div contenteditable>|foo</div>',
    'indent',
    [
        '<div contenteditable>',
            '<blockquote style="margin: 0 0 0 40px; border: none; padding: 0px;">',
                '|foo',
            '</blockquote>',
        '</div>'
    ].join('')),
    'Indent at caret');
</script>
